agents/requirements.txt
fastapi
uvicorn[standard]
anthropic
hedera-agent-kit>=3.2.0
hiero-sdk-python>=0.1.9
python-dotenv
pydantic
httpx
mypy
langchain>=1.0.0
langchain-anthropic
langchain-core
langgraph
web3
